chore: Move ConversationNewDeviceSystemMessageCellDescription into its own file
